Moved App.js to components folder
